Move Footer into Layout file to match others
