+++ src/index.js
@@ -1,7 +1,6 @@
 import MLButton from './ml-button'
 import MLCollapse from './ml-collapse'
 import MLDatePicker from './ml-date-picker'
-import MLFooter from './ml-footer'
 import MLInputNumber from './ml-input-number'
 import MLLayout from './ml-layout'
 import MLRadio from './ml-radio'
@@ -11,7 +10,6 @@ export {
   MLButton,
   MLCollapse,
   MLDatePicker,
-  MLFooter,
   MLInputNumber,
   MLLayout,
   MLRadio,
