+++ src/ml-layout.js
@@ -1,9 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { Layout } from 'antd'
-import MLFooter from './ml-footer'
-
-const { Header, Content, Sider } = Layout
+const { Header, Content, Sider, Footer } = Layout
 
 const layoutStyle = {}
 
@@ -13,7 +11,7 @@ const headerStyle = {
   alignItems: 'center',
   justifyContent: 'center',
   textAlign: 'center',
-  height: '64px'
+  height: '64px',
   // fontSize: 'TODO',
   // fontFamily: 'TODO',
 }
@@ -23,7 +21,7 @@ const siderStyle = {
   flexDirection: 'column',
   alignItems: 'center',
   justifyContent: 'center',
-  backgroundColor: '#dddddd'
+  backgroundColor: '#dddddd',
 }
 
 const contentStyle = {}
@@ -45,7 +43,7 @@ const MLHeader = (props) => {
 }
 
 const siderProps = {
-  width: 70
+  width: 70,
 }
 
 const MLSider = (props) => {
@@ -62,6 +60,58 @@ const MLContent = (props) => {
       {props.children}
     </Content>
   )
+}
+
+const footerStyleBase = {
+  padding: '12px 50px',
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+  flexDirection: 'column',
+  width: '100%',
+  textAlign: 'center',
+  margin: '0',
+  fontSize: '14px',
+  fontFamily: 'Helvetica Neue',
+}
+
+const footerStyleWhiteBg = Object.assign({}, footerStyleBase, {
+  color: '#999',
+  backgroundColor: '#fff',
+})
+
+const footerStyleGraphicBg = Object.assign({}, footerStyleBase, {
+  color: '#fff',
+  background: 'unset',
+})
+
+const footerSpanStyle = {
+  padding: '0 5px',
+}
+
+const MLFooter = (props) => {
+  const footerStyle = props.graphicBackground ? footerStyleGraphicBg : footerStyleWhiteBg
+  return (
+    <Footer style={footerStyle} {...props}>
+      <div>
+        <span style={footerSpanStyle}>Copyright @ {props.year} MarkLogic Corporation. All Rights Reserved.</span>
+        |
+        <span style={footerSpanStyle}>
+          <a href='#TODO'>Terms and Conditions</a>
+        </span>
+        |
+        <span style={footerSpanStyle}>
+          <a href='#TODO'>Policies</a>
+        </span>
+      </div>
+    </Footer>
+  )
+}
+
+// Typechecking for Ant Design properties
+MLFooter.propTypes = {
+  year: PropTypes.number,
+  graphicBackground: PropTypes.bool,
 }
 
 MLLayout.MLHeader = MLHeader
